requirements.txt
-i https://pypi.org/simple
backports.tarfile==1.2.0; python_version < '3.12'
certifi==2025.1.31; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
colorama==0.4.6; platform_system == 'Windows'
docutils==0.21.2; python_version >= '3.9'
flask==2.2.2; python_version >= '3.7'
flask-cors==5.0.0
greenlet==3.1.1; python_version >= '3.7'
gunicorn==23.0.0; python_version >= '3.7'
id==1.5.0; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
importlib-metadata==8.6.1; python_version < '3.12'
itsdangerous==2.2.0; python_version >= '3.8'
jaraco.classes==3.4.0; python_version >= '3.8'
jaraco.context==6.0.1; python_version >= '3.8'
jaraco.functools==4.1.0; python_version >= '3.8'
jinja2==3.1.5; python_version >= '3.7'
keyring==25.6.0; platform_machine != 'ppc64le' and platform_machine != 's390x'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
mdurl==0.1.2; python_version >= '3.7'
more-itertools==10.6.0; python_version >= '3.9'
nh3==0.2.20; python_version >= '3.8'
numpy==1.24.1; python_version >= '3.8'
packaging==24.2; python_version >= '3.8'
pandas==1.5.3; python_version >= '3.8'
pygments==2.19.1; python_version >= '3.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2025.1
pywin32-ctypes==0.2.3; sys_platform == 'win32'
readme-renderer==44.0; python_version >= '3.9'
requests==2.32.3; python_version >= '3.8'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rfc3986==2.0.0; python_version >= '3.7'
rich==13.9.4; python_full_version >= '3.8.0'
scipy==1.15.1; python_version >= '3.10'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlalchemy==2.0.1; python_version >= '3.7'
texas-hold-em-utils==0.1.7; python_version >= '3.8'
twine==6.1.0; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.3.0; python_version >= '3.9'
waitress==3.0.2; python_full_version >= '3.9.0'
werkzeug==2.2.2; python_version >= '3.7'
zipp==3.21.0; python_version >= '3.9'
